agent/requirements.txt
pystray
pillow
pyinstaller
